requirements.txt
Django==2.1.2
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
gunicorn==19.9.0
python-decouple==3.1
whitenoise==4.1.2
dj-database-url==0.5.0
